+++ lib/qc/modules/qc_tester/calculator/providers/calculator_scanner_provider.dart
@@ -1,10 +1,10 @@
 import 'dart:async';
 
 import 'package:core/core.dart';
-import 'package:core_widgets/core_widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_trc/qc/modules/qc_tester/calculator/resources/calculator_service.dart';
 import 'package:flutter_trc/qc/modules/qc_tester/calculator/resources/my_calculator_response.dart';
+import 'package:flutter_trc/qc/modules/qc_tester/lob_devices/resources/brand_list_response.dart';
 import 'package:provider/provider.dart';
 
 class CalculatorScannerProvider extends CalculatorServiceInitProvider {
@@ -21,6 +21,29 @@ class CalculatorScannerProvider extends CalculatorServiceInitProvider {
       } else {
         completer.completeError("Some internal error occurred");
       }
+    }, onError: (error) {
+      completer.completeError(ApiErrorHelper.getErrorMessage(error).toString());
+    });
+    return completer.future;
+  }
+
+  Future<List<BrandListData>> getBrandList(int categoryId) {
+    var completer = Completer<List<BrandListData>>();
+    service.getBrandList(categoryId).listen((event) {
+      completer.complete(event?.brandList);
+      notifyListeners();
+    }, onError: (error) {
+      completer.completeError(ApiErrorHelper.getErrorMessage(error).toString());
+    });
+    return completer.future;
+  }
+
+  Future<String> getCategory(String deviceBarcode) {
+    // TODO: need to configure this api
+    var completer = Completer<String>();
+    service.getCategory(deviceBarcode).listen((event) {
+      completer.complete("Success");
+      notifyListeners();
     }, onError: (error) {
       completer.completeError(ApiErrorHelper.getErrorMessage(error).toString());
     });