+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # TRC Releases
 
-# v4.0.0+65 (XX-01-2025)
+# v4.0.0+66 (XX-01-2025)
 
 ### Feature:
 
 * `Flutter upgrade to 3.27`[(SCM-8571)](https://reglobe.atlassian.net/browse/SCM-8571)
+* `CDP testing flow for Laptop`[(SCM-8681)](https://reglobe.atlassian.net/browse/SCM-8681)
+
+
+# v3.5.6+65 (16-01-2025)
+
+### Hot fix:
+
+* `QC - False QR code Scan on colour screen`[(SCM-8752)](https://reglobe.atlassian.net/browse/SCM-8752)
 
 
 # v3.5.5+64 (15-01-2025)